euclidean & optimized & color space

+++ utilities.py
@@ -2,6 +2,7 @@ import math
 import pickle
 import re
 import os
+import time
 from multiprocessing import current_process, Process
 from threading import Thread
 from glob import glob
@@ -56,7 +57,9 @@ class DatabaseStructure(object):
         return glob(self.image_folder + '/*' + self.postfix)
 
     def remove_existing_files(self):
-        shutil.rmtree(self.folder)
+        if os.path.isdir(self.folder):
+            shutil.rmtree(self.folder)
+
 
 def get_database_structure(folder):
     return DatabaseStructure(folder)
@@ -72,72 +75,79 @@ def load(filename):
     with open(filename, 'rb') as file:
         return pickle.load(file)
 
-# FOR MULTIPROCESSING:
-# usage: items, small list executor
 
-def multithreading_(items, small_list_executor, info=None):
-    threads = []
-    num_of_items = len(items)
-    num_of_threads = settings.MAX_THREAD_PER_PROCESS
-    if num_of_threads == 1:
-        num_of_items_per_thread = num_of_items  # if num of threads is 1
-    elif num_of_threads > 1:
-        num_of_items_per_thread = settings.MIN_ITEMS_PER_THREAD
-        while num_of_threads > 1:
-            num_of_items_per_thread = math.ceil(num_of_items / num_of_threads)
-            if num_of_items_per_thread < settings.MIN_ITEMS_PER_THREAD:
-                num_of_threads -= 1
-                num_of_items_per_thread = num_of_items  # for condition not meet in next loop
-            else:
-                break
+is_first_print = True
+last_percent = None
+last_percent_time_left = None
+last_percent_print_time = None
+est_time_lefts = []
+
+# est time left is very unaccurate
+def print_progress(curr, total, curr_chunk=None, total_chunks=None):
+    global is_first_print
+    global last_percent
+    global last_percent_time_left
+    global last_percent_print_time
+    global est_time_lefts
+    curr_percent = math.floor(curr / total * 100)
+    curr_time = time.time()
+    if is_first_print:
+        est_time_left = float("inf")
+        is_first_print = False
+        last_percent_time_left = est_time_left
+        last_percent_print_time = curr_time
+    elif last_percent == curr_percent:
+        est_time_left = last_percent_time_left
     else:
-        raise ValueError('number of threads must be at least 1, please check settings.py')
+        bad_est_time_left = (curr_time - last_percent_print_time) / (curr_percent - last_percent) * (100 - curr_percent)
+        est_time_lefts.append(bad_est_time_left)
+        est_time_left = sum(est_time_lefts[-3:]) / (3 if len(est_time_lefts) >= 3 else len(est_time_lefts))
+        last_percent_time_left = est_time_left
+        last_percent_print_time = curr_time
 
-    if num_of_threads > num_of_items:
-        num_of_threads = num_of_items
+    last_percent = curr_percent
 
-    for thread_count in range(0, num_of_threads):
-        start = thread_count * num_of_items_per_thread
-        end = (thread_count + 1) * num_of_items_per_thread
-        items_for_this_thread = items[start:end]
-        threads.append(Thread(target=small_list_executor, args=(items_for_this_thread, info), daemon=True))
-    print('  |->', current_process().name, '=>', len(threads), 'threads =>', num_of_items, 'items')
-    for thread in threads:
-        thread.start()
-    for thread in threads:
-        thread.join()
+    print('\r >>> {0} / {1} => {2}% | Time Left est. {3:.2f}s'.format(curr, total, curr_percent, est_time_left), end='')
+
+    if curr_chunk and total_chunks:
+        print(' | {} / {}'.format(curr_chunk, total_chunks), end='')
 
 
-# breaks the given items into small lists in different process and threads,
-# pass these small lists to small_list_executor
-# results_saver is used to save results if any
-# raise ValueError if num of max threads specified in settings is < 1
-def multiprocessing_(items, small_list_executor, info=None):
-    print('Starting multiprocessing...')
-    processes = []
-    num_of_items = len(items)
-    num_of_processes = os.cpu_count()
-    num_of_items_per_process = settings.MIN_ITEMS_PER_THREAD
-    while num_of_processes > 1:
-        num_of_items_per_process = math.ceil(num_of_items / num_of_processes)
-        if num_of_items_per_process < settings.MIN_ITEMS_PER_THREAD:
-            num_of_processes -= 1
-            num_of_items_per_process = num_of_items  # if next condition failed, num_of_processes = 1
-        else:
-            break
+def print_done(time_taken):
+    print(' [ done ] => {0:.2f}s'.format(time_taken))
+    global is_first_print
+    global last_percent
+    global last_percent_time_left
+    global last_percent_print_time
+    global est_time_lefts
+    is_first_print = True
+    last_percent = None
+    last_percent_print_time = None
+    last_percent_time_left = None
+    est_time_lefts = []
 
-    if num_of_processes > num_of_items:
-        num_of_processes = num_of_items
+# def remove_empty(a_list):
+#     if not isinstance(a_list, list):
+#         return a_list
+#
+#     if not a_list:
+#         return False
+#
+#     result = []
+#     for item in a_list:
+#         item = remove_empty(item)
+#         if item:
+#             result.append(item)
+#     return result
 
-    for process_count in range(0, num_of_processes):
-        start = process_count * num_of_items_per_process
-        end = (process_count + 1) * num_of_items_per_process
-        items_for_this_process = items[start:end]
-        processes.append(
-            Process(target=multithreading_, args=(items_for_this_process, small_list_executor, info),
-                    daemon=True))
-    print('|--', current_process().name, '=>', len(processes), 'processes =>', num_of_items, 'items')
-    for process in processes:
-        process.start()
-    for process in processes:
-        process.join()
+
+def remove_empty(a_list):
+    if not isinstance(a_list, list):
+        return a_list
+    result = []
+    for item in a_list:
+        if item:
+            item = remove_empty(item)
+            if item:
+                result.append(item)
+    return result
